fix breaking changes; file structure
