+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import Main from './Main';
-import Login from './components/LoginSignup/LoginSignup';
+import Main from './views/Main/Main';
+import Login from './views/LoginSignup/LoginSignup';
 
 const App = () => {
   return (
