+++ frontend/src/components/ChangeLog/ChangeLog.css
@@ -1,0 +1,38 @@
+.drawer {
+  width: 400px;
+}
+
+.header {
+  height: 75px;
+  background-color: #3e654a;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  padding-left: 20px;
+}
+
+.searchBar {
+  width: 70%;
+  margin: 20px 50px;
+  display: flex;
+  align-items: center;
+  padding: 10px;
+  box-shadow: 0 2px 30px rgba(0, 0, 0, 0.1);
+}
+
+.searchIcon {
+  height: 100%;
+  position: absolute;
+  right: 15%;
+  pointer-events: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.selectMenu {
+  width: 45%;
+  margin: 20px 50px;
+  display: flex;
+  box-shadow: 0 2px 30px rgba(0, 0, 0, 0.1);
+}
