+++ frontend/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import Main from "./Main"
+import App from './App';
 import './index.css';
 
 createRoot(document.getElementById("root")).render(
-  <Main />
+  <App />
 )
